# perf: Quick prefix validation check