--- a/src/cli/add.rs
+++ b/src/cli/add.rs
@@ -9,7 +9,7 @@ use crate::{
     cli::cli_config::{DependencyConfig, PrefixUpdateConfig, ProjectConfig},
     environment::verify_prefix_location_unchanged,
     load_lock_file,
-    lock_file::{filter_lock_file, LockFileDerivedData, UpdateContext},
+    lock_file::{filter_lock_file, LockFileDerivedData, UpdateContext, UpdateMode},
     project::{grouped_environment::GroupedEnvironment, DependencyType, Project},
 };
 use clap::Parser;
@@ -293,7 +293,7 @@ pub async fn execute(args: Args) -> miette::Result<()> {
         && default_environment_is_affected
     {
         updated_lock_file
-            .prefix(&project.default_environment(), true)
+            .prefix(&project.default_environment(), UpdateMode::Force)
             .await?;
     }