--- a/src/cli/install.rs
+++ b/src/cli/install.rs
@@ -1,5 +1,6 @@
 use crate::cli::cli_config::ProjectConfig;
 use crate::environment::update_prefix;
+use crate::lock_file::UpdateMode;
 use crate::Project;
 use clap::Parser;
 use fancy_display::FancyDisplay;
@@ -52,7 +53,13 @@ pub async fn execute(args: Args) -> miette::Result<()> {
         let environment = project.environment_from_name_or_env_var(Some(env))?;
 
         // Update the prefix by installing all packages
-        update_prefix(&environment, args.lock_file_usage.into(), false, true).await?;
+        update_prefix(
+            &environment,
+            args.lock_file_usage.into(),
+            false,
+            UpdateMode::Force,
+        )
+        .await?;
 
         installed_envs.push(environment.name().clone());
     }