--- a/src/cli/project/channel/remove.rs
+++ b/src/cli/project/channel/remove.rs
@@ -1,3 +1,4 @@
+use crate::lock_file::UpdateMode;
 use crate::{
     environment::{update_prefix, LockFileUsage},
     Project,
@@ -16,7 +17,7 @@ pub async fn execute(mut project: Project, args: AddRemoveArgs) -> miette::Resul
         &project.default_environment(),
         LockFileUsage::Update,
         args.no_install,
-        true,
+        UpdateMode::Force,
     )
     .await?;
     project.save()?;