--- a/src/cli/project/platform/add.rs
+++ b/src/cli/project/platform/add.rs
@@ -48,6 +48,7 @@ pub async fn execute(mut project: Project, args: Args) -> miette::Result<()> {
         &project.default_environment(),
         LockFileUsage::Update,
         args.no_install,
+        true,
     )
     .await?;
     project.save()?;