--- a/src/cli/remove.rs
+++ b/src/cli/remove.rs
@@ -6,6 +6,7 @@ use crate::DependencyType;
 use crate::Project;
 
 use crate::cli::cli_config::{DependencyConfig, PrefixUpdateConfig, ProjectConfig};
+use crate::lock_file::UpdateMode;
 
 use super::has_specs::HasSpecs;
 
@@ -82,7 +83,7 @@ pub async fn execute(args: Args) -> miette::Result<()> {
             &project.default_environment(),
             prefix_update_config.lock_file_usage(),
             prefix_update_config.no_install,
-            true,
+            UpdateMode::Force,
         )
         .await?;
     }