--- a/src/cli/remove.rs
+++ b/src/cli/remove.rs
@@ -83,7 +83,7 @@ pub async fn execute(args: Args) -> miette::Result<()> {
             &project.default_environment(),
             prefix_update_config.lock_file_usage(),
             prefix_update_config.no_install,
-            UpdateMode::Force,
+            UpdateMode::Revalidate,
         )
         .await?;
     }