--- a/src/cli/run.rs
+++ b/src/cli/run.rs
@@ -5,7 +5,6 @@ use miette::{Diagnostic, IntoDiagnostic};
 use std::collections::hash_map::Entry;
 use std::collections::HashSet;
 use std::convert::identity;
-use std::time::Instant;
 use std::{collections::HashMap, string::String};
 
 use crate::cli::cli_config::{PrefixUpdateConfig, ProjectConfig};
@@ -82,7 +81,6 @@ pub async fn execute(args: Args) -> miette::Result<()> {
             .into_diagnostic()?;
     }
 
-    let start = Instant::now();
     // Ensure that the lock-file is up-to-date.
     let mut lock_file = project
         .update_lock_file(UpdateLockFileOptions {
@@ -90,8 +88,6 @@ pub async fn execute(args: Args) -> miette::Result<()> {
             ..UpdateLockFileOptions::default()
         })
         .await?;
-    let duration = start.elapsed();
-    println!("Lock file updated in {:?}", duration);
 
     // Construct a task graph from the input arguments
     let search_environment = SearchEnvironments::from_opt_env(