--- a/src/cli/shell.rs
+++ b/src/cli/shell.rs
@@ -9,6 +9,7 @@ use rattler_shell::{
 };
 
 use crate::cli::cli_config::{PrefixUpdateConfig, ProjectConfig};
+use crate::lock_file::UpdateMode;
 use crate::{
     activation::CurrentEnvVarBehavior, environment::update_prefix,
     project::virtual_packages::verify_current_platform_has_required_virtual_packages, prompt,
@@ -241,7 +242,7 @@ pub async fn execute(args: Args) -> miette::Result<()> {
         &environment,
         args.prefix_update_config.lock_file_usage(),
         false,
-        false,
+        UpdateMode::Validate,
     )
     .await?;