--- a/src/cli/shell.rs
+++ b/src/cli/shell.rs
@@ -242,7 +242,7 @@ pub async fn execute(args: Args) -> miette::Result<()> {
         &environment,
         args.prefix_update_config.lock_file_usage(),
         false,
-        UpdateMode::Validate,
+        UpdateMode::QuickValidate,
     )
     .await?;