--- a/src/cli/shell_hook.rs
+++ b/src/cli/shell_hook.rs
@@ -113,6 +113,7 @@ pub async fn execute(args: Args) -> miette::Result<()> {
         &environment,
         args.prefix_update_config.lock_file_usage(),
         false,
+        false,
     )
     .await?;