--- a/src/lib.rs
+++ b/src/lib.rs
@@ -5,7 +5,7 @@ pub mod environment;
 mod global;
 mod install_pypi;
 mod install_wheel;
-mod lock_file;
+pub mod lock_file;
 mod prefix;
 mod project;
 mod prompt;