… docs/advanced/explain_info_command.md
… docs/features/environment.md
… docs/reference/cli.md
… src/lock_file/mod.rs
… src/lock_file/records_by_name.rs
… src/lock_file/update.rs
… tbump.toml
… tests/common/mod.rs
… tests/install_tests.rs
… tests/integration/common.py
… tests/integration/test_run_cli.py